+++ connector/solarwindsentityconnector/factory.go
@@ -48,17 +48,18 @@ func createConnector(settings connector.Settings, config component.Config, logs 
 		return nil, fmt.Errorf("expected config of type *Config, got %T", config)
 	}
 	events := cfg.Schema.NewEvents(settings.TelemetrySettings)
+	attributeMapper := internal.NewAttributeMapper(cfg.Schema.NewEntities())
 
 	se := &solarwindsentity{
 		logger: settings.Logger,
 		eventDetector: internal.NewEventDetector(
-			cfg.Schema.NewEntities(),
-			cfg.SourcePrefix,
-			cfg.DestinationPrefix,
+			attributeMapper,
 			events.LogEvents,
 			events.MetricEvents,
 			settings.Logger,
 		),
+		sourcePrefix:     cfg.SourcePrefix,
+		destPrefix:       cfg.DestinationPrefix,
 		expirationPolicy: cfg.Expiration,
 		logsConsumer:     logs,
 	}