+++ connector/solarwindsentityconnector/internal/relationship.go
@@ -20,16 +20,11 @@ import (
 	"time"
 )
 
-type RelationshipEntity struct {
-	Type string
-	IDs  pcommon.Map
-}
-
 type Relationship struct {
 	Action      string
 	Type        string
-	Source      RelationshipEntity
-	Destination RelationshipEntity
+	Source      Entity
+	Destination Entity
 	Attributes  pcommon.Map
 }
 
@@ -44,7 +39,7 @@ var _ Event = (*Relationship)(nil)
 //   - Relationship type
 //   - timestamp
 //   - event type set to update action
-func (r *Relationship) Update(logRecords *plog.LogRecordSlice) {
+func (r Relationship) Update(logRecords *plog.LogRecordSlice) {
 	logRecord := logRecords.AppendEmpty()
 	logRecord.SetObservedTimestamp(pcommon.NewTimestampFromTime(time.Now()))
 	attrs := logRecord.Attributes()
@@ -75,7 +70,7 @@ func (r *Relationship) Update(logRecords *plog.LogRecordSlice) {
 //   - Relationship type
 //   - timestamp
 //   - event type set to delete action
-func (r *Relationship) Delete(logRecords *plog.LogRecordSlice) {
+func (r Relationship) Delete(logRecords *plog.LogRecordSlice) {
 	logRecord := logRecords.AppendEmpty()
 	logRecord.SetObservedTimestamp(pcommon.NewTimestampFromTime(time.Now()))
 	attrs := logRecord.Attributes()