+++ connector/solarwindsentityconnector/internal/storage/internal_storage_test.go
@@ -35,7 +35,7 @@ var cfg = &config.ExpirationSettings{
 	TTLCleanupIntervalSeconds: ttlCleanupInterval,
 }
 
-var sourceEntity = internal.RelationshipEntity{
+var sourceEntity = internal.Entity{
 	Type: "service",
 	IDs: func() pcommon.Map {
 		m := pcommon.NewMap()
@@ -44,7 +44,7 @@ var sourceEntity = internal.RelationshipEntity{
 	}(),
 }
 
-var destEntity = internal.RelationshipEntity{
+var destEntity = internal.Entity{
 	Type: "database",
 	IDs: func() pcommon.Map {
 		m := pcommon.NewMap()
@@ -252,7 +252,7 @@ func TestTtlExpiration_TenDifferentUpdates_ResultInTenExpiryEvents(t *testing.T)
 
 	for i := 0; i < 10; i++ {
 		// Create a unique relationship by varying the source entity's ID
-		sourceEntity := internal.RelationshipEntity{
+		sourceEntity := internal.Entity{
 			Type: "service",
 			IDs: func() pcommon.Map {
 				m := pcommon.NewMap()
@@ -325,7 +325,7 @@ func TestTtlExpiration_RelationshipIsRemovedFirst_EntitiesSecond(t *testing.T) {
 	storage.entities.Wait()
 	storage.relationships.Wait()
 
-	kb := NewKeyBuilder()
+	kb := internal.NewKeyBuilder()
 	sourceHash, err := kb.BuildEntityKey(relationship.Source)
 	require.NoError(t, err)
 	destHash, err := kb.BuildEntityKey(relationship.Destination)
@@ -414,7 +414,7 @@ func TestInternalStorage_Delete(t *testing.T) {
 			name: "Error building relationship hash key",
 			relationship: &internal.Relationship{
 				Type: "", // Empty type should cause an error when building the key
-				Source: internal.RelationshipEntity{
+				Source: internal.Entity{
 					Type: "service",
 					IDs:  pcommon.NewMap(),
 				},
@@ -532,7 +532,7 @@ func TestDelete_ComparedToTTLExpiration(t *testing.T) {
 	createTestRelationship := func(suffix string) *internal.Relationship {
 		return &internal.Relationship{
 			Type: "testRelation" + suffix,
-			Source: internal.RelationshipEntity{
+			Source: internal.Entity{
 				Type: "sourceType",
 				IDs: func() pcommon.Map {
 					m := pcommon.NewMap()
@@ -540,7 +540,7 @@ func TestDelete_ComparedToTTLExpiration(t *testing.T) {
 					return m
 				}(),
 			},
-			Destination: internal.RelationshipEntity{
+			Destination: internal.Entity{
 				Type: "destType",
 				IDs: func() pcommon.Map {
 					m := pcommon.NewMap()