+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## vNext
-- `solarwindsentityconnector` Fix extra output log issue by adding missing yaml tags for cache configuration parsing
+- `solarwindsentityconnector` Added support for prefixes on entities when created from relationships.
+- `solarwindsentityconnector` Fix extra output log issue by adding missing yaml tags for cache configuration parsing.
 
 ## v0.127.5
 - `solarwindsentityconnector` Benchmark fix